+++ data-raw/maya-rda/Bridgelux/make-bridgelux-mspct.R
@@ -33,6 +33,7 @@ names(what.measured) <- new.names
 bridgelux.mspct <- source_mspct()
 for (s in spectra) {
   temp.spct <- get(s)
+  temp.spct <- setNormalised(temp.spct)
   temp.spct <- normalize(temp.spct)
   temp.spct <- smooth_spct(temp.spct)
   temp.spct <- thin_wl(temp.spct)