+++ data-raw/maya-rda/Epileds/make-epileds-mspct.R
@@ -29,6 +29,7 @@ for (s in spectra) {
   }
   cat(s, "\n")
   temp.spct <- get(s)
+  temp.spct <- setNormalised(temp.spct)
   temp.spct <- normalize(temp.spct)
   temp.spct <- smooth_spct(temp.spct)
   temp.spct <- thin_wl(temp.spct)