+++ data-raw/maya-rda/Epileds/make-more-epileds-mspct.R
@@ -28,6 +28,7 @@ epileds.mspct <- source_mspct()
 for (s in spectra) {
   cat(s, "\n")
   temp.spct <- get(s)
+  temp.spct <- setNormalised(temp.spct)
   temp.spct <- normalize(temp.spct)
   temp.spct <- smooth_spct(temp.spct)
   temp.spct <- thin_wl(temp.spct)