+++ data-raw/maya-rda/Ledguhon/make-ledguhon-mspct.R
@@ -29,6 +29,7 @@ ledguhon.mspct <- source_mspct()
 for (s in spectra) {
   temp.spct <- get(s)
   temp.spct <- fshift(temp.spct, c(320, 350))
+  temp.spct <- setNormalised(temp.spct)
   temp.spct <- normalize(temp.spct)
   temp.spct <- smooth_spct(temp.spct)
   temp.spct <- thin_wl(temp.spct)