+++ data-raw/maya-rda/Nichia/make-nichia-mspct.R
@@ -75,6 +75,7 @@ for (s in type2name.map) {
                         "from Lumitronix (https://www.lumitronix.com/en_gb/) ca. 2015-2019.")
   what.measured <- paste("LED type", name2type.map[s], "from Nichia")
   temp.spct <- get(s)
+  temp.spct <- setNormalised(temp.spct)
   temp.spct <- normalize(temp.spct)
   temp.spct <- smooth_spct(temp.spct)
   temp.spct <- thin_wl(temp.spct)