+++ data-raw/maya-rda/Seoul/make-seoul-mspct.R
@@ -67,6 +67,7 @@ for (s in type2name.map) {
                         "supplier:", supplier[name2type.map[s]])
   what.measured <- paste("LED type", name2type.map[s], "from Seoul Semiconductor")
   temp.spct <- get(s)
+  temp.spct <- setNormalised(temp.spct)
   temp.spct <- normalize(temp.spct, norm = "max", unit.out = "energy")
   temp.spct <- smooth_spct(temp.spct)
   temp.spct <- thin_wl(temp.spct)